rename old dashit directory to dashit_filter, update setuptools
